refactor encoded bid checks into auctions util file with other nft related calls
